% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covplant.R
\docType{data}
\name{covplant}
\alias{covplant}
\title{Covariate data}
\format{An object of class \code{data.frame} with 2366 rows and 3 columns.}
\source{
no source
}
\usage{
covplant
}
\description{
covplant is the dataset of the spatial covariate of the plantillness dataset. Each column can be used as a covariate along the plantillness dataset in the estima function. Each row of a selected covariate matches a point of the plantillness dataset, that's why there is no coordinates in the covplant dataset.
}
\details{
\itemize{
\item $v1 spatial covariate
\item $v2 spatial covariate
\item $v3 spatial covariate
}
}
\keyword{dataset}
\keyword{datasets}
