% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantillness.R
\docType{data}
\name{plantillness}
\alias{plantillness}
\title{Illness data}
\format{An object of class \code{data.frame} with 2366 rows and 16 columns.}
\source{
no source
}
\usage{
plantillness
}
\description{
plantillness is a dataset representing a vineyard of the Bordeaux region
}
\details{
\itemize{
\item $NRang : rows of the vines
\item $NCep : position in the row of the vines
\item $X2004 : state of the vine at the position (NRang,NCep) in 2004
\item $X2005 : state of the vine at the position (NRang,NCep) in 2005
\item $X2006 : state of the vine at the position (NRang,NCep) in 2006
\item $X2007 : state of the vine at the position (NRang,NCep) in 2007
\item $X2008 : state of the vine at the position (NRang,NCep) in 2008
\item $X2009 : state of the vine at the position (NRang,NCep) in 2009
\item $X2010 : state of the vine at the position (NRang,NCep) in 2010
\item $X2011 : state of the vine at the position (NRang,NCep) in 2011
\item $X2012 : state of the vine at the position (NRang,NCep) in 2012
\item $X2013 : state of the vine at the position (NRang,NCep) in 2013
\item $X2014 : state of the vine at the position (NRang,NCep) in 2014
\item $X2015 : state of the vine at the position (NRang,NCep) in 2015
\item $X2016 : state of the vine at the position (NRang,NCep) in 2016
\item $X2017 : state of the vine at the position (NRang,NCep) in 2017
}
}
\keyword{dataset}
\keyword{datasets}
